re-add zig bindings
Updated to be compatible zig v0.14.0

Changes:
1. use the master branch of zig-tree-sitter
2. remove the exported language() function

In (1), the latest release of zig-tree-sitter does not yet
use the enum literal for the package name, but the recent commits
fixed it, so use the master branch for now until they make a new
release.

The problem with (2) is there are conflicting module versions
when tree-sitter-c and zig-tree-sitter are both added as a dependency
in a project. This is because zig-tree-sitter also imports its
own version of tree-sitter-c.

So instead of importing tree-sitter-c directly as such:

  const language = @import("tree-sitter-c").language();

tree-sitter-c can just be linked during build, and
declare an extern reference:

  extern fn tree_sitter_c() callconv(.C) *ts.Language;
  // ...
  const language = tree_sitter_c();